--- src/app/components/browser/mod.rs
@@ -1,5 +1,0 @@
-mod browser;
-mod browser_model;
-
-pub use browser::*;
-pub use browser_model::*;
