+++ src/app/components/library/mod.rs
@@ -1,0 +1,5 @@
+mod library;
+mod library_model;
+
+pub use library::*;
+pub use library_model::*;
