+++ src/app/components/navigation/mod.rs
@@ -1,7 +1,10 @@
 mod navigation;
-mod navigation_model;
-
 pub use navigation::*;
+
+mod navigation_model;
 pub use navigation_model::*;
 
 mod home;
+
+mod factory;
+pub use factory::*;
