+++ src/app/components/now_playing/mod.rs
@@ -1,2 +1,5 @@
 mod now_playing;
 pub use now_playing::*;
+
+mod now_playing_model;
+pub use now_playing_model::*;
