+++ src/app/components/playlist/mod.rs
@@ -1,5 +1,5 @@
 mod playlist;
-mod playlist_model;
-
 pub use playlist::*;
-pub use playlist_model::*;
+
+mod song;
+pub use song::*;
