--- src/app/components/song/mod.rs
@@ -1,2 +1,0 @@
-mod song;
-pub use song::*;
