+++ src/meson.build
@@ -34,47 +34,48 @@ sources = files(
 './app/components/details/details.rs',
 './app/components/details/details_model.rs',
 './app/components/details/mod.rs',
-'./app/components/playback/playback_model.rs',
-'./app/components/playback/playback.rs',
+'./app/components/utils.rs',
+'./app/components/playback/playback_info.rs',
+'./app/components/playback/playback_control.rs',
 './app/components/playback/mod.rs',
 './app/components/album/album.rs',
 './app/components/album/mod.rs',
 './app/components/artist_details/artist_details.rs',
 './app/components/artist_details/artist_details_model.rs',
 './app/components/artist_details/mod.rs',
-'./app/components/browser/browser_model.rs',
-'./app/components/browser/browser.rs',
-'./app/components/browser/mod.rs',
-'./app/components/song/song.rs',
-'./app/components/song/mod.rs',
 './app/components/login/login.rs',
 './app/components/login/mod.rs',
 './app/components/login/login_model.rs',
+'./app/components/library/library_model.rs',
+'./app/components/library/library.rs',
+'./app/components/library/mod.rs',
 './app/components/player_notifier.rs',
 './app/components/now_playing/now_playing.rs',
 './app/components/now_playing/mod.rs',
+'./app/components/now_playing/now_playing_model.rs',
 './app/components/user_menu/user_menu_model.rs',
 './app/components/user_menu/mod.rs',
 './app/components/user_menu/user_menu.rs',
+'./app/components/playlist/song.rs',
 './app/components/playlist/playlist.rs',
-'./app/components/playlist/playlist_model.rs',
 './app/components/playlist/mod.rs',
 './app/components/navigation/navigation.rs',
+'./app/components/navigation/factory.rs',
 './app/components/navigation/home.rs',
 './app/components/navigation/mod.rs',
 './app/components/navigation/navigation_model.rs',
+'./app/components/notification/mod.rs',
 './app/components/search/search_bar_model.rs',
 './app/components/search/search_model.rs',
 './app/components/search/search_bar.rs',
 './app/components/search/mod.rs',
 './app/components/search/search.rs',
 './app/components/mod.rs',
-'./app/components/notification.rs',
+'./app/loader.rs',
+'./app/list_store.rs',
 './app/dbus/mpris.rs',
 './app/dbus/mod.rs',
 './app/dbus/types.rs',
-'./app/loader.rs',
-'./app/list_store.rs',
 './app/credentials.rs',
 './app/models.rs',
 './app/backend/cache.rs',
