+++ src/spot.gresource.xml
@@ -12,12 +12,12 @@
     <file alias="components/details.ui">app/components/details/details.ui</file>
     <!-- artist details -->
     <file alias="components/artist_details.ui">app/components/artist_details/artist_details.ui</file>
-    <!-- browser/library -->
-    <file alias="components/browser.ui">app/components/browser/browser.ui</file>
+    <!-- library -->
+    <file alias="components/library.ui">app/components/library/library.ui</file>
     <!-- now playing -->
     <file alias="components/now_playing.ui">app/components/now_playing/now_playing.ui</file>
     <!-- song -->
-    <file alias="components/song.css">app/components/song/song.css</file>
-    <file alias="components/song.ui">app/components/song/song.ui</file>
+    <file alias="components/song.css">app/components/playlist/song.css</file>
+    <file alias="components/song.ui">app/components/playlist/song.ui</file>
   </gresource>
 </gresources>
