+++ src/window.ui
@@ -216,6 +216,7 @@
                     </child>
                     <child>
                       <object class="GtkLabel" id="current_song_info">
+                        <property name="width-request">90</property>
                         <property name="visible">True</property>
                         <property name="can-focus">False</property>
                         <property name="halign">start</property>
